requirements_dev.txt
pip==19.2.3
bump2version==0.5.11
wheel==0.33.6
watchdog==0.9.0
tox==3.14.0
coverage==4.5.4
Sphinx==7.2.6
twine==5.0.0
ruff==0.3.5
torch==2.2.1
torchvision==0.17.1
Pillow>=9.0.0
numpy>=1.20.0
